backend/requirements.admin.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-jose==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0 